+++ src/py/aggregators/focidataset.py
@@ -1,6 +1,7 @@
 import hashlib
 import os.path
 import pickle
+import random
 import time
 from typing import Dict, List
 
@@ -14,14 +15,18 @@ from src.sammie.py.SessionData import SessionData
 from src.sammie.py.modules.ModuleBase import ModuleBase
 from src.sammie.py.util import shapeutil
 
+CUR_FILE_VERSION:int = 2
 
 def __getBatchHash(batchKey: List[str]):
     return hashlib.md5(
         ''.join(batchKey).encode()).hexdigest()  # The hash used for storage a simple concatenation of batchKeys
 
 
-def __getBaseAggregateFile():
-    return {'data': {}, 'created': time.time(), 'batchinfo': {}}
+def __getBaseAggregateFile(ts:float = 0):
+    if ts == 0: ts = time.time()
+
+    #Version 1.1 was released together with HARLEY version 1.2.3
+    return {'data': {}, 'created': ts, 'batchinfo': {}, 'version':CUR_FILE_VERSION, 'log':{}}
 
 
 def __getDataSetInfoObject(curData) -> AggregatorFileInfo:
@@ -29,7 +34,7 @@ def __getDataSetInfoObject(curData) -> AggregatorFileInfo:
         return AggregatorFileInfo(True, True, 'File exists, but contains no results yet.')
     batchInfoArr = [curData['batchinfo'][c] for c in curData['batchinfo']]
     return AggregatorFileInfo(True, True, 'File contains results from %d batches' % (len(curData['data'])),
-                              batchInfoArr)
+                              batchInfoArr,curData['log'])
 
 
 def __getBatchData(session: SessionData, modulesById: Dict[str, ModuleBase], scalePxToNm: float = None):
@@ -48,6 +53,30 @@ def __getBatchData(session: SessionData, modulesById: Dict[str, ModuleBase], sca
     # We store the source image and the contours the user accepted in this dataset.
     return CellsDataset(denoisedImage, acceptedContourList, scalePxToNm, maskFile.ref, imgShift)
 
+def __updateOldFileVersion(curData):
+    oldVersion = 1
+    if 'version' in curData:
+        oldVersion = curData['version']
+
+    if oldVersion == CUR_FILE_VERSION:
+        return curData, False
+
+    newData = __getBaseAggregateFile(curData['created'])
+    upgraded = False
+    if(oldVersion == 1): #update to 2
+        #Key Change
+        upgraded = True
+        seed = random.randint(0,2**32)
+        for oldBatchNum in curData['data']:
+            #Create a random hash key
+            bk = hashlib.md5(('%d%d'%(seed,oldBatchNum)).encode()).hexdigest()
+            newData['data'][bk] = curData['data'][oldBatchNum]
+            newData['batchinfo'][bk] = AggregatorBatchInfo([bk],curData['created'],'Missing source paths, due to file upgrade')
+
+        newData['log']['%s|warning'%str(time.time())] = 'This file was upgraded from version 1 to 2. The main change was the addition of source file paths to identify batches. This info is not available for batches stored before the upgrade and the filepaths therefore will be displayed as a unique random string. The data and functionality of this file are not affected by the change.'
+        newData['version'] = 2
+
+    return newData, upgraded
 
 def appendToCellSet_Info(destinationPath: str) -> AggregatorFileInfo:
     filename = os.path.basename(destinationPath)
@@ -66,6 +95,13 @@ def appendToCellSet_Info(destinationPath: str) -> AggregatorFileInfo:
     try:
         with open(destinationPath, 'rb') as handle:
             curData = pickle.load(handle)
+            curData, upgraded = __updateOldFileVersion(curData)
+
+        if upgraded: #If file version has been upgraded, save the file
+            with open(destinationPath, 'wb') as handle:
+                pickle.dump(curData, handle, protocol=pickle.HIGHEST_PROTOCOL)
+
+
     except:
         return AggregatorFileInfo(True, True,
                                   'File exists but is corrupt and can\'t be loaded. It will be overwritten on export.')